requirements.txt
msgraph-sdk
langchain
openai
python-dotenv
pymilvus
tiktoken
